Work around Eiger bug when using Internal/External Enable mode
There is a bug in the FileWriter that creates nimages*ntriggers data links
inside the master file when on Internal/External Enable modes, but only
'ntriggers' data files will be created.

This commit sets nimages to 1 when on one of these modes and then sets it back
to its previous value when done.