Change doc example trigger mode to External Series
An example refers to an acquisition with TriggerMode set to "External
Enable", but that mode only generates one image per trigger pulse, which
does not make sense in the example.  Change the TriggerMode to "External
Series", which generates N images per trigger pulse and is the intended
TriggerMode for the example.